fix issue with std::sort, fix SOAContainer comparisons, cleanups
The std::sort issue was caused by sort storing individual contained
objects outside the container (e.g. when moving lots of elements
around, for example when moving a single element to later position in
the array while preserving the relative order of the other elements);
this means that the value_type must be the underlying std::tuple, and
not the proxy type. This (in turn) requires the SKIN to forward copy and
move assignments to the underlying proxy. Once that is fixed, std::sort
works as advertised.

The "fallout" from this is that the sort functor itself changes
semantics a little, since it can no longer rely on the skinned proxy's
interface, but only has std::tuple to work with. Since sorting isn't
that frequent, that's an acceptable drawback for now, and has been
documented appropriately.

In the future, it may be possible to extend the SKIN to cover the raw
value_type as well (without too much boilerplate code), and that
annoyance would be gone as well...