SOAView: initial hackings
SOAView, the new class helping to implement composibility, starts out as
a straight copy of SOAContainer. The idea is to slowly morph it into
something that'll handle the range-based interface part of SOAContainer,
i.e. operations that do not change the size of the range; this includes
operations such as:

- size(), max_size(), empty()
- [cr]{begin,end}()
- front(), back()
- operator[](), at()
- operator=(), assign()
- swap

SOAContainer could then inherit from SOAView, thus reusing code, and
adding the remaining operations that you can perform on a proper
container...

As a result of this refactoring, SOAView will need somewhat better
checking of its template arguments, since there are more possibilities
for callers/users to screw up inadvertently. This commit implements a
first shot at such an improved verification.

Another thing that is planned is to detect if any of the contained
fields is const, and if so, force all the iterators and references
const. That'll help detect errors in user code. The rule is simple: If
you want a SOAView to be modifiable, make sure none of the fields are
const... (That's so obvious it should go without saying, right? ;)