more work towards convenient fields
I think I've seen more of Florian's deep wisdom: The field should be a
class, not a template. The "rebind" has moved into a subclass of the
field called and accessor, and the SkinBase and skin creation macros
have been updated accordingly. Added a couple of convenience macros so
users don't have to type quite so much...