take care of an issue exposed by gcc 7.2
Tuples don't convert back and forth so freely any more, so we needed a
slight change in SOADressedTuple. Moreover, there was an issue in the
new smart skins when used in conjunction with std::sort. This was
resolved in SkinBase and FieldsBase, mostly, and should also bring
along a speedup in compilation time, since we now cache the field
indices in the typelist once a concrete skin is instantiated.