enhance compatibility with compilers and C++ standard libraries
After upgrading to gcc 7.2, it turns out that the code no longer
compiled because the C++ standard library is a lot stricter about
which types a tuple can be constructed or assigned from. Clang is also
stricter about this, and libc++ is a lot stricter than early versions
of libstdc++. SOA::DressedTuple has been adapted to compiled with
libstdc++ and gcc versions 4.8, 4.9, 5.5. 6.3, 7.2, and clang versions
3.8, 3.9, 4.0, 5.0, 6.0, and with libc++ and clang versions 3.8, 3.9,
4.0, 5.0, 6.0.

No functional changes, though.