examples/Points.cc: make examples safe
pushing back into a SOA::Container that's based on tuple<float, float>
a tuple<double, double> is a type mismatch, and therefore it is a bad
idea to do that in the first place (especially if the skin has custom
constructors). Besides, if you want to construct the contained objects
with a call to push_back/emplace_back, emplace_back is the way to go
(and parameters convert as expected in the call to the constructor).

Let's not go there, and the clang 3.8/libc++ issue is gone, too