fix a use-reference-after-object-out-of-scope bug in zip
Thanks to Paul Seyfert for reporting and writing a test.

Basically, the issue was that neither SOA::view nor SOA::zip were fully
safe against rvalue references. For SOA::view, we have these cases:

- SOA::view<...>(view)

  depending on the argument view and the fields selected, some of the
  fields may have their underlying storage moved (if view is an rvalue
  reference)

- SOA::view<...>(view, first, last)

  this is never owning, since it's a subrange, and is therefore never
  moving anything

SOA::zip inherited the defects of SOA::view, and created a few of its
own due to a bad implementation.

This commit is a complete rewrite of both SOA::view and SOA::zip, where
special attention was given to make it rvalue-safe (i.e. to make sure
that all routines move the underlying storage when it's legal and you'd
expect them to).