fix compiler warnings, make older compilers happy
Apart from a signed-unsigned comparison, this mostly makes older C++
compilers happy which seem to choke on the more complex templated
forwarding constructors in SkinBase (and also on a couple of other
classes in the inheritance hierarchy). The solution is to bite the sour
apple and default the usual set of default, copy and move constructor
and also the copy and move and assignment operators. A comment was added
where this was done.