first shot at a transform-like algorithm that uses tagged types
The new code should make writing transform-like loops much easier, since
the selection of fields can now be done based on the argument list of
the functor/function passed to transform. Output is to a SOA::Container
with fields based on the return type of the functor/function. This used
tagged types extensively. Example:

using namespace SOA;
auto view = /* from somewhere */;
auto cont = transform(view, [] (value<field_1> x, cref<field_2> y,
				const LHCb::State& s) {
    return std::make_tuple(value<field_sum>(x + y),
			   value<field_z>(s.z()));
    });
// cont is now a container with two fields: field_sum and field_z

In the example above, tagged types (value<field>, ref<field>,
cref<field>) are used by default, but if the field can be uniquely
determined by the untagged type (the LHCb::State above), that is
permissible as well. (The code will produce an error message if the
argument list of the functor cannot be matched one-to-one and
unabiguously to the fields present in the view that is passed.)
If only a single field is returned, it is not necessary to put it inside
a std::tuple, i.e. "return value<field>(42);" is fine.

It's possible to give custom skins or use something other than
std::vector for the underlying storage of the container - see the
documentation