Revert "Try and avoid gcc bug."
This reverts commit 2fed1a2c79660637ee405eda31b97114e86eac86.

As this appears to be a gcc bug (and a rather "fragile" one at that
which is sensitive to changing the string lengths by single characters,
for example), I won't apply this "fix". Instead we're going to move to
std::ostringstream instead of std::stringstream (which I should have
done from the start).