try and avoid what appears to be a gcc bug
This bug is triggered in the unit tests of the LHCb nightlies (if you
compile them in a directory whose path has the right length). As it is a
rather fragile bug in gcc, this change happens to "fix" the issue as
well, and it's a change that should have been in from the start.

Note that until gcc is fixed, there is no guarantee that some other
piece of production code won't trigger the same bug again.